// ui/src/app/pages/ingest/models/dbconnectors.ts
export interface Connectors {
    pipelineName: string;
    connectorName: string;
    hostName: string;
    port: number;
    database: string;
    userName: string;
    password: any;
    databaseName: string;
    serviceName: string;
    connectorType: string;
    saveDbCon: boolean;
    projectId: number;
    userId: number;
    schemaName: string;
    tableName: string;
    pipelineId: number;

}